v0.0.4
Version 0.0.4: Many formatting changes and completed first stage of nation data display. Changed the way battle comparisons are done for Social Temperature (still figuring out what this should mean). Fixed a few inconsistencies. Fixed issue where nation foreground and background colors could match making them unreadable.

+++ BadAIWarz/Nation.cs
@@ -10,6 +10,6 @@ public class Nation
     public int MilitaryStrength { get; set; }
     public int TerritorySize { get; set; }
     public double EconomyWorth { get; set; }
-    public int SocialTemperature { get; set; }
+    public double SocialTemperature { get; set; }
     public int PopulationSizeAdult { get; set; }
 }
